Created an rnn base class that the lstm models and the dag transformer extend to avoid code duplication
